+++ R/Get_Registry.R
@@ -109,7 +109,13 @@ Get_Registry <- function(Year = 2023, filename = c("SCUANAGRAFESTAT", "SCUANAAUT
         status <- 0
       }
       if(status != 200){
-        message("Operation exited with status: ", status, "; operation repeated")
+        attempt <- attempt + 1
+        message("Operation exited with status: ", status, "; operation repeated (",
+                10 - attempt, " attempts left)")
+      }
+      if(attempt >= 10) {
+        message("Maximum attempts reached. Abort. We apologise for the inconvenience")
+        return(NULL)
       }
     }