+++ R/Get_School2mun.R
@@ -77,15 +77,22 @@ Get_School2mun <- function(Year = 2023, show_col_types = FALSE, verbose = TRUE,
   home.url <-"https://dati.istruzione.it/opendata/opendata/catalogo/elements1/?area=Edilizia%20Scolastica"
   homepage <- NULL
   attempt <- 0
-  while(is.null(homepage) && attempt <= 10){
+  while(is.null(homepage)){
     homepage <- tryCatch({
       xml2::read_html(home.url)
     }, error = function(e){
-      message("Cannot read the html; ", 10 - attempt,
-              " attempts left. If the problem persists, please contact the mantainer.\n")
+      message("Cannot read the html. If the problem persists, please contact the mantainer.\n")
       return(NULL)
     })
-    attempt <- attempt + 1
+    if(status != 200){
+      attempt <- attempt + 1
+      message("Operation exited with status: ", status, "; operation repeated (",
+              10 - attempt, " attempts left)")
+    }
+    if(attempt >= 10) {
+      message("Maximum attempts reached. Abort. We apologise for the inconvenience")
+      return(NULL)
+    }
   }
   if(is.null(homepage)) return(NULL)
   name_pattern <- "([0-9]+)\\.(csv)$"